Move Signer from jwt into types
