+++ src/types.rs
@@ -1,11 +1,17 @@
-use std::fmt;
 use std::sync::Arc;
+use std::{fmt, io};
 
 use hyper::Client;
 use hyper_rustls::HttpsConnectorBuilder;
+use ring::{
+    rand::SystemRandom,
+    signature::{RsaKeyPair, RSA_PKCS1_SHA256},
+};
 use serde::Deserializer;
 use serde::{Deserialize, Serialize};
 use time::{Duration, OffsetDateTime};
+
+use crate::Error;
 
 /// Represents an access token. All access tokens are Bearer tokens.
 ///
@@ -67,6 +73,52 @@ impl Token {
     /// Get expiry of token, if available
     pub fn expires_at(&self) -> Option<OffsetDateTime> {
         self.inner.expires_at
+    }
+}
+
+/// A JSON Web Token ready for signing.
+pub(crate) struct Signer {
+    key: RsaKeyPair,
+    rng: SystemRandom,
+}
+
+impl Signer {
+    pub(crate) fn new(pem_pkcs8: &str) -> Result<Self, Error> {
+        let private_keys = rustls_pemfile::pkcs8_private_keys(&mut pem_pkcs8.as_bytes());
+
+        let key = match private_keys {
+            Ok(mut keys) if !keys.is_empty() => {
+                keys.truncate(1);
+                keys.remove(0)
+            }
+            Ok(_) => {
+                return Err(io::Error::new(
+                    io::ErrorKind::InvalidInput,
+                    "Not enough private keys in PEM",
+                )
+                .into())
+            }
+            Err(_) => {
+                return Err(io::Error::new(
+                    io::ErrorKind::InvalidInput,
+                    "Error reading key from PEM",
+                )
+                .into())
+            }
+        };
+
+        Ok(Signer {
+            key: RsaKeyPair::from_pkcs8(&key).map_err(|_| Error::SignerInit)?,
+            rng: SystemRandom::new(),
+        })
+    }
+
+    pub(crate) fn sign(&self, input: &[u8]) -> Result<Vec<u8>, Error> {
+        let mut signature = vec![0; self.key.public_modulus_len()];
+        self.key
+            .sign(&RSA_PKCS1_SHA256, &self.rng, input, &mut signature)
+            .map_err(|_| Error::SignerFailed)?;
+        Ok(signature)
     }
 }
